+++ protoc_plugin/lib/message_generator.dart
@@ -523,6 +523,19 @@ class MessageGenerator extends ProtobufContainer {
                 fieldPathSegment: memberFieldPath,
                 start: 'void '.length)
           ]);
+      if (field.baseType.isMessage) {
+        out.printlnAnnotated(
+            '${fieldTypeString} ${names.ensureMethodName}() => \$_ensure('
+            '${field.index}, '
+            '${field.number}, '
+            '${fieldTypeString}.create);',
+            <NamedLocation>[
+              NamedLocation(
+                  name: names.ensureMethodName,
+                  fieldPathSegment: memberFieldPath,
+                  start: '${fieldTypeString} '.length)
+            ]);
+      }
     }
   }