+++ protoc_plugin/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 19.0.0
+
+* Breaking: GeneratedMessage classes now have methods `ensureX` for each message field X.
+Generates code that requires at least `protobuf` 0.14.3.
+
 ## 18.0.1
 
 * Add a `bin/protoc-gen-dart.bat` script making it easier to compile on windows using a local